backend/requirements.txt
fastapi
uvicorn[standard]
pandas
python-multipart
google-genai
python-dotenv
httpx
websockets
passlib[bcrypt]
python-jose[cryptography]
sqlalchemy
